protosanitizer: log messages without secrets
When running at glog level >= 5, CSI sidecar apps log the full
CreateVolumeRequest, including the secrets. Secrets should never be
logged at any level to avoid accidentally exposing them.

We need to filter out the secrets. With older CSI versions, that could
have been done based on the field name, which is still an option
should this get backported. With CSI 1.0, a custom field option marks
fields as secret. Using that option has the advantage that the code
will continue to work also when new secret fields get added in the
future.

For the sake of simplicity, JSON is now used as representation of the
string instead of the former compact text format from gRPC. That makes
it possible to strip values from a map with generic types, instead of
having to copy and manipulate the real generated structures.

Another option would have been to copy
https://github.com/golang/protobuf/blob/master/proto/text.go and
modify it so that it skips secret fields, but that's over 800 lines of
code.

This version of the code is identical to the one reviewed in
https://github.com/kubernetes-csi/external-provisioner/pull/171 with
the backwards parameters in
assert.NotContains (https://github.com/kubernetes-csi/external-provisioner/pull/171#pullrequestreview-179404145)
fixed.