protosanitizer: handle secrets in oneof and maps
As expected, oneof wasn't supported so far and requires special code:
- JSON naming is not defined in Go structs and thus inconsistent with
  protobuf
- JSON and Go struct are nested, protobuf fields aren't

For maps from plain type to a message, the messages were not filtered.

Because the nesting got a bit deep with support for both added, the
code now avoids nesting by aborting the flow in `if` cases with
continue or return and continuing unindented after that.

Everything that could cause a panic gets checked, which protects
against malformed protobuf data but should never trigger in
practice. To determine where an assumption gets violated, strip() now
returns an error and wraps that with pkg.errors, which eventually gets
reports in the resulting string. This is similar to how fmt.Sprintf
handles parameter errors.