protosanitizer: add test for converted message
One real use case is that users of an older spec can receive a message
based on a new spec (possible as long as backwards compatibility is
preserved) without logging any secrets added in the new spec. It is
unlikely, but it is nonetheless better to have a test for it.