protosanitizer: remove upstream dependency on CSI 1.0
We cannot depend on the github.com/container-storage-interface/spec
because then the package cannot be vendored into other repos where an
older revision is needed, like the 0.4.x sidecars.

We only really need the "csi_secrets" extension which has to be fixed
from now on, therefore we can simply copy csi.pb.go manually once and
then use it with a different package name, without having to worry
about updating it.