protosanitizer: remove dependency on csi package
The only thing that is needed from the spec is the extension field
definition. It's easier to understand why the code is generic and
going to support future revisions of the spec when using a copy of
that field definition instead of the entire spec package.

Another positive effect is that only a single file needs to be
vendored and compiled by users of the code and not also the more
complex spec.