connection: handle connection loss
Some but not all CSI sidecar apps may want to cache information
retrieved from a CSI driver. Those (and only) those apps needs a way
to detect that the cached information became invalid due to a CSI
driver restart and then react to that. The Connect function supports
notifying the app via an optional callback which then can:
- kill the app via os.Exit
- invalidate the cached information and continue by reconnecting
- prevent reconnecting and exit the application when gRPC method
  calls fail with status.Unavailable

The default behavior is as before, i.e. the connection is getting
restablished.