replace dep with "go mod"
This follows the industry trend.

A side effect is that consumers of this package might now be able to
select a newer version of the Kubernetes dependencies (k8s.io/api,
k8s.io/apimachinery, k8s.io/client-go).

The exact version for those used to be selected in Gopkg.toml and
because of the lack of semantic versioning of Kubernetes packages, dep
then enforced using exactly those versions also in consumers of the
package. We couldn't express that any version >= 1.14 is okay.

We still can't express that with go.mod. With go mod, consumers now
may have more flexibility, but may also accidentally choose older,
incompatible versions. With dep, the situation is the same due to the
lack of a Gopk.toml file.