metrics: support usage inside CSI driver
A CSI driver is a gRPC server, which implies that the interceptor must
use a slightly different API. The extended CSIMetricsManager gets
exported with a suitable method for that.

To avoid counting the same operation twice in the same metric, CSI
sidecar and driver should use different subsystem names. Two new
functions provide the CSIMetricsManager that they are expected to use
for the sake of consistency.

However, special cases need additional flexibility:
- constant labels (same for all samples)
- varying labels (same label names, but per-sample values)
- configurable subsystem
- configurable stability