connection: exit with less output on connection loss
klog.Fatal dumps information about all running goroutines when the
connection to the CSI driver is lost. Loosing the connection is normal
and depends on the order in which containers are shut down. If it
happens, then the reason is unlikely to be related to goroutines.

Therefore this extra output is not helpful or worse, fills up logfiles
when a sidecar has many goroutines, as in the external-provisioner
which runs 100 workers by default.