[rpc] Run the probe first
The previous beheavior of ProbeForever was to try a probe and then sleep for the `probeInterval` duration. When it was modified to use a `time.Ticker` instead of a sleep, it changed the behavior to wait the `probeInterval` first and then do the probe. This commit restore the previous behavior by running the probe first. Based on https://stackoverflow.com/a/54752803.